Made the pagign test more useful
Added a method to VirtualMemoryManager to convert a virtual address to a physical one

Removed  s_KernelStart and s_KernelEnd from kernel, they werent doing what i expected and turns out they werent even needed

removed kernel.h as it was only storing the above two variables.

+++ kernel/arch/i386/kernel/MemoryManagement/PageDirectoryEntry.cpp
@@ -19,31 +19,31 @@ namespace kernel::memory
 
 	void PageDirectoryEntry::SetFrame(physical_addr addr)
 	{
-		m_Entry = (m_Entry & ~Flags::FRAME) | addr;
+		m_Entry = (m_Entry & ~(uint32_t)Flags::FRAME) | addr;
 	}
 
 	bool PageDirectoryEntry::IsPresent()
 	{
-		return m_Entry & Flags::PRESENT;
+		return m_Entry & (uint32_t)Flags::PRESENT;
 	}
 
 	bool PageDirectoryEntry::IsWritable()
 	{
-		return m_Entry & Flags::WRITABLE;
+		return m_Entry & (uint32_t)Flags::WRITABLE;
 	}
 
 	physical_addr PageDirectoryEntry::GetFrameAddress()
 	{
-		return m_Entry & Flags::FRAME;
+		return m_Entry & (uint32_t)Flags::FRAME;
 	}
 
 	bool PageDirectoryEntry::IsUserMode()
 	{
-		return m_Entry & Flags::USER;
+		return m_Entry & (uint32_t)Flags::USER;
 	}
 
 	bool PageDirectoryEntry::Uses4MBPages()
 	{
-		return m_Entry & Flags::FOURMB;
+		return m_Entry & (uint32_t)Flags::FOURMB;
 	}
 }
